moved LongRangeQuery in LongRangeQuery.cs and added tests
